[2/10] Use float instead of double for color calculations
Similar to the previous patch, but now for video/ instead of console/.

The console/ code mostly only had addition and multiplication floating point
operations. Here we also have some square-root and power operations (in the
initialization code, not on the fast code paths). Especially those latter
operations are more expensive (I mean double vs float) in a soft-float
implementation.

This patch (mostly) only changes color calculations (gamma, brightness, ...).
Other double-calculations are left for later.

+++ src/video/FBPostProcessor.cc
@@ -208,7 +208,7 @@ void FBPostProcessor<Pixel>::drawNoiseLine(
 template <class Pixel>
 void FBPostProcessor<Pixel>::drawNoise(OutputSurface& output)
 {
-	if (renderSettings.getNoise() == 0.0) return;
+	if (renderSettings.getNoise() == 0.0f) return;
 
 	unsigned height = output.getHeight();
 	unsigned width = output.getWidth();
@@ -310,8 +310,8 @@ void FBPostProcessor<Pixel>::paint(OutputSurface& output)
 		//fprintf(stderr, "post processing lines %d-%d: %d\n",
 		//	srcStartY, srcEndY, lineWidth );
 		output.lock();
-		double horStretch = renderSettings.getHorizontalStretch();
-		unsigned inWidth = unsigned(horStretch + 0.5);
+		float horStretch = renderSettings.getHorizontalStretch();
+		unsigned inWidth = unsigned(horStretch + 0.5f);
 		std::unique_ptr<ScalerOutput<Pixel>> dst(
 			StretchScalerOutputFactory<Pixel>::create(
 				output, pixelOps, inWidth));
